requirements.txt
# Multi-Agent Sales Data Analysis System

# Agent Framework
openai-agents[litellm]>=0.1.0

# LLM Clients
openai>=1.93.0
anthropic>=0.76.0

# Data Processing
pandas>=2.0.0
numpy>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Forecasting
statsmodels>=0.14.0
# prophet  # Optional - install separately: pip install prophet

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Frontend
gradio>=5.0.0

# Database
pymongo>=4.6.0
motor>=3.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
